python/requirements.txt
fastapi
uvicorn[standard]
pyyaml
jinja2
aiofiles
pyngrok
pandas 
matplotlib 
seaborn
